Adding unit test for last reparent function.
